+++ push_swap/libraries/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ibenaiss <ibenaiss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 18:46:14 by ibenaiss          #+#    #+#             */
-/*   Updated: 2024/03/24 16:13:34 by ibenaiss         ###   ########.fr       */
+/*   Updated: 2024/03/25 14:38:45 by ibenaiss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,39 +22,39 @@
 # include <string.h>
 # include <unistd.h>
 
-int		ft_isalpha(const int c);
-char	*ft_strchr(const char *str, int c);
-int		ft_isdigit(int arg);
-int		ft_isalnum(const int c);
-int		ft_isascii(int c);
-int		ft_isprint(int c);
-size_t	ft_strlen(const char *str);
-void	*ft_memset(void *str, int c, size_t n);
-void	ft_bzero(void *ptr, size_t size);
-void	*ft_memcpy(void *dest, const void *src, size_t n);
-void	*ft_memmove(void *dest, const void *src, size_t n);
-size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-size_t	ft_strlcat(char *dest, const char *src, size_t size);
-long	ft_atoi(const char *nbr);
-int		ft_toupper(int c);
-char	*ft_strrchr(const char *str, int c);
-int		ft_tolower(int c);
-void	*ft_memchr(const void *s, int c, size_t n);
-int		ft_memcmp(const void *str1, const void *str2, size_t n);
-char	*ft_strnstr(const char *big, const char *little, size_t len);
-char	*ft_strdup(const char *str);
-void	ft_bzero(void *ptr, size_t size);
-void	*ft_calloc(size_t size, size_t count);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-char	*ft_strtrim(char const *s1, char const *set);
-char	*ft_itoa(int n);
-void	ft_putchar_fd(char c, int fd);
-void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(int n, int fd);
-void	ft_putstr_fd(char *s, int fd);
-char	**ft_split(char const *s, char c);
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-void	ft_striteri(char *s, void (*f)(unsigned int, char *));
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
+int			ft_isalpha(const int c);
+char		*ft_strchr(const char *str, int c);
+int			ft_isdigit(int arg);
+int			ft_isalnum(const int c);
+int			ft_isascii(int c);
+int			ft_isprint(int c);
+size_t		ft_strlen(const char *str);
+void		*ft_memset(void *str, int c, size_t n);
+void		ft_bzero(void *ptr, size_t size);
+void		*ft_memcpy(void *dest, const void *src, size_t n);
+void		*ft_memmove(void *dest, const void *src, size_t n);
+size_t		ft_strlcpy(char *dest, const char *src, size_t size);
+size_t		ft_strlcat(char *dest, const char *src, size_t size);
+long double	ft_atoi(const char *nbr);
+int			ft_toupper(int c);
+char		*ft_strrchr(const char *str, int c);
+int			ft_tolower(int c);
+void		*ft_memchr(const void *s, int c, size_t n);
+int			ft_memcmp(const void *str1, const void *str2, size_t n);
+char		*ft_strnstr(const char *big, const char *little, size_t len);
+char		*ft_strdup(const char *str);
+void		ft_bzero(void *ptr, size_t size);
+void		*ft_calloc(size_t size, size_t count);
+char		*ft_substr(char const *s, unsigned int start, size_t len);
+char		*ft_strtrim(char const *s1, char const *set);
+char		*ft_itoa(int n);
+void		ft_putchar_fd(char c, int fd);
+void		ft_putendl_fd(char *s, int fd);
+void		ft_putnbr_fd(int n, int fd);
+void		ft_putstr_fd(char *s, int fd);
+char		**ft_split(char const *s, char c);
+char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+void		ft_striteri(char *s, void (*f)(unsigned int, char *));
+int			ft_strncmp(const char *s1, const char *s2, size_t n);
 
 #endif
